Change database to SQLite
The JSON "database" has grown to 300 MB and it's becoming inconvenient
to keep it all in memory every time we need to do anything.

Use SQLite instead, and Sequel as a light ORM layer, because I just
hate seeing "insert into ... values (?, ?, ?, ?, ?, ?, ?, ?)" and
"where foo in (?, ?, ?, ?, ?, ?, ?, ?)".

Some of the code still loops over everything instead of doing nice
queries, because I got bored rewriting it.

Define dependencies using Bundler, and install them using
`bundle install --standalone`. This is needed because the Ruby
environment on Toolforge is hopeless, I'm dealing with 3 different
Ruby versions and 3 different sets of gems depending on whether a
script runs from the webservice, a job or interactive session.